+++ app/Views/layouts/sidebar-karyawan.php
@@ -88,7 +88,6 @@
                 <a href="<?= url('/karyawan/leave/create') ?>" class="flex items-center px-2 py-1.5 text-body rounded-base hover:bg-neutral-tertiary hover:text-fg-brand group">
                     <i class="fa-solid fa-calendar-day transition duration-75 group-hover:text-fg-brand"></i>
                     <span class="flex-1 ms-3 whitespace-nowrap">Penjauan cuti</span>
-                    <span class="inline-flex items-center justify-center w-4.5 h-4.5 ms-2 text-xs font-medium text-fg-danger-strong bg-danger-soft border border-danger-subtle rounded-full">2</span>
                 </a>
             </li>
             <li>